requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
asgiref==3.3.1
    # via django
atomicwrites==1.4.0
    # via pytest
attrs==20.3.0
    # via pytest
certifi==2020.12.5
    # via requests
chardet==3.0.4
    # via requests
colorama==0.4.4
    # via pytest
django-filter==2.4.0
    # via -r requirements.in
django==3.0.5
    # via
    #   -r requirements.in
    #   django-filter
    #   djangorestframework
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==4.3.0
    # via -r requirements.in
djangorestframework==3.11.0
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
gunicorn==20.0.4
    # via -r requirements.in
idna==2.10
    # via requests
iniconfig==1.1.1
    # via pytest
packaging==20.9
    # via pytest
pluggy==0.13.1
    # via pytest
psycopg2-binary==2.8.5
    # via -r requirements.in
py==1.10.0
    # via pytest
pyjwt==2.0.1
    # via djangorestframework-simplejwt
pyparsing==2.4.7
    # via packaging
pytest-django==3.10.0
    # via -r requirements.in
pytest==6.2.3
    # via pytest-django
pytz==2021.1
    # via django
requests==2.23.0
    # via -r requirements.in
six==1.15.0
    # via -r requirements.in
sqlparse==0.4.1
    # via django
toml==0.10.2
    # via pytest
urllib3==1.25.11
    # via requests

# The following packages are considered to be unsafe in a requirements file:
# setuptools
